# docs/api/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Outlier API",
  "description": "유튜브 다채널 유해댓글 관리 서비스",
  "version": "0.1.0"
 },
 "paths": {
...
  "/api/channels/{channel_id}/rules": {
   "get": {
    "tags": [
     "rules"
    ],
    "summary": "등록된 단어 목록",
    "operationId": "list_rules_api_channels__channel_id__rules_get",
    "parameters": [
     {
      "name": "channel_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Channel Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/RuleOut"
         },
         "title": "Response List Rules Api Channels  Channel Id  Rules Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "post": {
    "tags": [
     "rules"
    ],
    "summary": "단어 등록",
    "operationId": "create_rule_api_channels__channel_id__rules_post",
    "parameters": [
     {
      "name": "channel_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Channel Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/RuleCreate"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/RuleOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/channels/{channel_id}/rules/{rule_id}": {
   "patch": {
    "tags": [
     "rules"
    ],
    "summary": "단어 수정",
    "operationId": "update_rule_api_channels__channel_id__rules__rule_id__patch",
    "parameters": [
     {
      "name": "channel_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Channel Id"
      }
     },
     {
      "name": "rule_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Rule Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/RuleUpdate"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/RuleOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "delete": {
    "tags": [
     "rules"
    ],
    "summary": "단어 삭제",
    "operationId": "delete_rule_api_channels__channel_id__rules__rule_id__delete",
    "parameters": [
     {
      "name": "channel_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Channel Id"
      }
     },
     {
      "name": "rule_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Rule Id"
      }
     }
    ],
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "RuleCreate": {
    "properties": {
     "rule_value": {
      "type": "string",
      "maxLength": 255,
      "minLength": 1,
      "title": "Rule Value",
      "description": "등록할 단어",
      "examples": [
       "시발"
      ]
     },
     "action": {
      "type": "string",
      "enum": [
       "block",
       "review",
       "allow"
      ],
      "title": "Action",
      "description": "block=즉시 차단 / review=보류 / allow=예외",
      "default": "block"
     },
     "expand_variants": {
      "type": "boolean",
      "title": "Expand Variants",
      "description": "초성·공백·특수문자 변형까지 잡을지",
      "default": true
     }
    },
    "type": "object",
    "required": [
     "rule_value"
    ],
    "title": "RuleCreate"
   },
   "RuleOut": {
    "properties": {
     "id": {
      "type": "integer",
      "title": "Id"
     },
     "channel_id": {
      "type": "integer",
      "title": "Channel Id"
     },
     "rule_value": {
      "type": "string",
      "title": "Rule Value"
     },
     "action": {
      "type": "string",
      "enum": [
       "block",
       "review",
       "allow"
      ],
      "title": "Action"
     },
     "expand_variants": {
      "type": "boolean",
      "title": "Expand Variants"
     },
     "enabled": {
      "type": "boolean",
      "title": "Enabled"
     },
     "compiled_regex": {
      "type": "string",
      "title": "Compiled Regex"
     },
     "created_at": {
      "type": "string",
      "format": "date-time",
      "title": "Created At"
     }
    },
    "type": "object",
    "required": [
     "id",
     "channel_id",
     "rule_value",
     "action",
     "expand_variants",
     "enabled",
     "compiled_regex",
     "created_at"
    ],
    "title": "RuleOut"
   },
   "RuleUpdate": {
    "properties": {
     "action": {
      "anyOf": [
       {
        "type": "string",
        "enum": [
         "block",
         "review",
         "allow"
        ]
       },
       {
        "type": "null"
       }
      ],
      "title": "Action"
     },
     "expand_variants": {
      "anyOf": [
       {
        "type": "boolean"
       },
       {
        "type": "null"
       }
      ],
      "title": "Expand Variants"
     },
     "enabled": {
      "anyOf": [
       {
        "type": "boolean"
       },
       {
        "type": "null"
       }
      ],
      "title": "Enabled"
     }
    },
    "type": "object",
    "title": "RuleUpdate"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   }
  }
 }
}